+++ examples/PFHub1a.cpp
@@ -1,5 +1,5 @@
-#include <CommandLine.hpp>
-#include <PFHub.hpp>
+#include <CabanaPF_CommandLine.hpp>
+#include <CabanaPF_PFHub.hpp>
 
 using namespace CabanaPF;
